deep_research/requirements.txt
fastapi
uvicorn[standard]
google-genai
psycopg2-binary
streamlit
requests
python-multipart
python-dotenv
matplotlib
pandas
google-generativeai 
filelock 
openai 
pydantic 
